review: the fence wait's helper lives with the link, a pending flag, a private cell
